Cast signal and total tag calculations to int64 for improved accuracy in GlobalBackgroundFit

+++ src/hotspot3/background_fit/fit.py
@@ -216,8 +216,8 @@ class GlobalBackgroundFit(BackgroundFit):
         )
         best_fit_result.n_signal = signal_mask.sum()
 
-        best_fit_result.signal_tags = agg_cutcounts[signal_mask].compressed().sum()
-        best_fit_result.total_tags = agg_cutcounts.compressed().sum()
+        best_fit_result.signal_tags = agg_cutcounts[signal_mask].compressed().astype(np.int64).sum()
+        best_fit_result.total_tags = agg_cutcounts.compressed().astype(np.int64).sum()
 
         return best_fit_result
